+++ tests/e2e/utils/src/flows/with-rest-api.js
@@ -1,24 +1,23 @@
 import factories from '../factories';
-import {Coupon, Setting, SimpleProduct} from '@woocommerce/api';
+import {Coupon, Setting, SimpleProduct, Order} from '@woocommerce/api';
 
 const client = factories.api.withDefaultPermalinks;
 const onboardingProfileEndpoint = '/wc-admin/onboarding/profile';
 const shippingZoneEndpoint = '/wc/v3/shipping/zones';
 const shippingClassesEndpoint = '/wc/v3/products/shipping_classes';
 const userEndpoint = '/wp/v2/users';
-const ordersEndpoint = '/wc/v3/orders';
 
 /**
  * Utility function to delete all merchant created data store objects.
  *
  * @param repository
  * @param defaultObjectId
+ * @param statuses Status of the object to check
  * @returns {Promise<void>}
  */
-const deleteAllRepositoryObjects = async ( repository, defaultObjectId = null ) => {
+ const deleteAllRepositoryObjects = async ( repository, defaultObjectId = null, statuses = [ 'draft', 'publish', 'trash' ] ) => {
 	let objects;
 	const minimum = defaultObjectId == null ? 0 : 1;
-	const statuses = ['draft','publish','trash'];
 
 	for ( let s = 0; s < statuses.length; s++ ) {
 		const status = statuses[ s ];
@@ -91,15 +90,8 @@ export const withRestApi = {
 	deleteAllOrders: async () => {
 		// We need to specfically filter on order status here to make sure we catch all orders to delete.
 		const orderStatuses = ['pending', 'processing', 'on-hold', 'completed', 'cancelled', 'refunded', 'failed', 'trash'];
-		for (let s = 0; s < orderStatuses.length; s++) {
-			const orders = await client.get( ordersEndpoint + `?status=${orderStatuses[s]}` );
-			if ( orders.data && orders.data.length ) {
-				for ( let o = 0; o < orders.data.length; o++ ) {
-					const response = await client.delete( ordersEndpoint + `/${orders.data[o].id}?force=true` );
-					expect( response.statusCode ).toBe( 200 );
-				}
-			}
-		}
+		const repository = Order.restRepository( client );
+		await deleteAllRepositoryObjects( repository, null, orderStatuses );
 	},
 	/**
 	 * Use api package to delete shipping zones.